pyproject.toml
[project]
name = "diploma"
version = "0.1.0"
description = "InsideOut: LangGraph-based emotional intelligence framework (CU Diploma)"
requires-python = ">=3.12"
dependencies = [
    "beartype>=0.22.9",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "nltk>=3.9.3",
    "python-dotenv>=1.2.2",
    "rouge-score>=0.1.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter (replaces flake8, isort, black)
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line too long — handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"src/run_erc.py" = ["E402"]  # load_dotenv before src imports is intentional
"src/run_erg.py" = ["E402"]
"tests/*.py" = ["E402"]  # module docstring before imports is standard in test files

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Pyright — static type checker
# ---------------------------------------------------------------------------

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
reportMissingModuleSource = false
reportMissingTypeStubs = false
# LangGraph ships incomplete stubs — these fire only on its internal types
reportUnknownMemberType = false
reportUnknownArgumentType = false
